add shift and caps lock keys and improved arrows

+++ includes/utils.h
@@ -10,7 +10,7 @@ uint8_t		vga_entry_color(enum vga_color fg, enum vga_color bg);
 uint16_t	vga_entry(unsigned char uc, uint8_t color);
 size_t		strlen(const char* str);
 uint8_t		inb(uint16_t port);
-uint8_t		keyboard_get_scancode();
+uint16_t	keyboard_get_scancode();
 void		outb(uint16_t port, uint8_t value);
 
 #endif
